First design for Searching screens options
I added the first rough design for the Searching screens options, these options will allow the user to select an algorithm, generate an array, add an element to the array, delete an element from the array, clear an array, control the speed of the algorithm and start the chosen algorithm (No functionality added yet). The button to start the algorithm has been positioned at the bottom of the frame, this creates a gap to add any extra options that might be needed later. A new class called SharedLayout was created, this is because the Sorted screen and Searching screen will have similar layouts, this new class groups together all the similar aspects of the two screens designs, so code is written several time (reducing the number of line).