Improved options design for Searching screen
I added a new option to allow the user to choose what the algorithms would search for. I also made the options nicer and more intuitive to look at by rearranging the layout, adding more spacing between widgets and keeping the blank space for extra options. I made it so that textboxes contain some default text telling the user what the purpose of the textboxes are to make them more clearer. The default text disappears on click, this only happens with the default text, user input is not cleared when they click on the textbox. I added some functionality, allowing elements to be added to an array, for now they can only add integers (but multiple different types can easily be allowed).