Added function to add numbers to the array
The user can now enter a number which will be displayed on the screen. The array can also never exceed a certain point but the centering is off slightly (I don't think this is something I can fix). In addtion, the searching screen works dynamically with different widths and heights.