Added deleteElement() and clearArray() methods
The deleteElement method deletes the last element from the integer array and also deletes the associated label from the screen. The clearArray method removes everything from the integer array and off the screen. Both deleteElement and clearArray remove the comma added to each element from the last element, so the removeComma function was written to handle this