Replaced number visualisation with bars
Due to issues with centring I decided to replace my intial idea of using just numbers, replacing each number with a bar proportional to the elements value. Functions where added to calculate the largest number that could be represented on screen and the amount of numbers the array can contain. I also wrote a function that calculates the amount of space needed to centre the displayed array.