Rewrote add(), delete() and clear() methods
The add(), delete() and clear() methods have been rewritten to support the use of bars. calculatePadding() was renamed to calculateSmallestPadding() and a new function called calculatePadding() was created using the old name. calculatePadding() calculates the padding needed to centre a displayed array of any size as best as possible. randomAdd() was renamed to randomGenerate(), this new name is clearer and more descriptive. Since add() and randomGenerate() both need to draw the array, the for loop used to draw the array in randomGenerate() was moved into it's own function called displayArray() that both add() and randomGenerate() call.