Reduced amount frames needed in Searching screen
The number of frames needed to create the Searching screen was decreased, this was done by making the algorithm info frame the same height as the home button frame, this was done as I have decided to not continue with the idea to outline an algorithms steps using text, instead the algorithm info frame will just display the time and space complexity of a given algorithm. I removed the add, target, delete and clear options, they will be replaced with a single slider to increase and decrease the arrays size. The reason I will do this is that it wasn't really clear what values the user can and can't add to the array the slider will remove any confusion, making the UI easier to understand