Added changing array size with scroll bar
The array can now shrink and grow in size by dragging the given scroll bar left and right. This is far quicker than pressing a button to generate an array of random size and manually adding or deleting elements to get the desired size. Users now have direct control over the arrays size.