Automation for when a new algorithm is added
A new package called searching_algorithms has been made, it will contain all the searching algorithm related files. The algorithm.py file contains an abstract class, which contains a method each algorithm must implenent (this is important for automatically adding each algorithm to the drop down box). The algorithm_names.py file contains the code for the automation, it gets every file in the package and filters only the module names, each module is dynamically imported and a function is called to return the algorithms name which is added to a list, this list of algorithm names is converted to a tuple before being passed back to the drop down menu and displayed to the user