Added Visualiation of Linear Search
I'm back after 3 months, I was too busy with work over the summer but now that university has started up I can continue development. I added a new file called init_algorithm.py, it contains a function called callAgorithm, when the users presses the solve button this function creates an instance of the relevant algorithm class and allows the algorithm to actually run. I rewrote some functions in algorithm_names.py to better filter out any files that don't follow the naming coventions and any potential classes that don't follow the specified format (pending markdown update). I then wrote a linear search that actually works! I also set values for each speed so the user can control how fast algorithms run through the array. Finally I added a new slider that lets the user choose how 'random' the target the algorithm looks for is, the user can decide if the target is in the array, if the target is not in that array or if they want the program to just generate a target.